+++ castle/cms/interfaces/controlpanel.py
@@ -7,6 +7,7 @@ from zope import schema
 from zope.interface import Interface
 from zope.schema.vocabulary import SimpleVocabulary
 
+
 def create_term(val, label):
     return SimpleVocabulary.createTerm(val, val, label)
 
@@ -176,7 +177,6 @@ class IAnnouncementData(Interface):
 
 
 class ISiteConfiguration(Interface):
-
     image_repo_location = schema.TextLine(
         title=u'Image repo path',
         required=True,
@@ -379,9 +379,9 @@ class IArchivalSettings(Interface):
 
 
 class ICastleSettings(
-        ISiteConfiguration,
-        IAPISettings,
-        IArchivalSettings):
+    ISiteConfiguration,
+    IAPISettings,
+    IArchivalSettings):
     pass