+++ HISTORY.rst
@@ -5,7 +5,14 @@ Changelog
 2.0.46 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Changed collective.documentviewer dep. version
+  [obct537]
+
+- Added status control panel to give users the status of relevant subprocesses
+  [mattjhess]
+
+- in query listing tile, do not display event start/end if they don't exist
+  [tkimnguyen]
 
 
 2.0.45 (2018-07-13)